+++ karma.conf.js
@@ -3,9 +3,15 @@
 
 module.exports = function(config) {
   var customLaunchers = {
-    sl_safari: {
+    sl_safari_8: {
       base: 'SauceLabs',
-      browserName: 'safari'
+      browserName: 'safari',
+      version: '8'
+    },
+    sl_safari_9: {
+      base: 'SauceLabs',
+      browserName: 'safari',
+      version: '9'
     },
     sl_chrome: {
       base: 'SauceLabs',
@@ -14,11 +20,6 @@ module.exports = function(config) {
     sl_firefox: {
       base: 'SauceLabs',
       browserName: 'firefox'
-    },
-    sl_ie_10: {
-      base: 'SauceLabs',
-      browserName: 'internet explorer',
-      version: '10'
     },
     sl_ie_11: {
       base: 'SauceLabs',